check: Fail blocked checker after check interval
Previously we use to warn about blocked checker every check interval.
However when a blocked checker returned, the check failed only if the
underlying dd process failed. This could cause the storage domain to be
valid while reading from storage blocked for long time. If a VM was
paused during this time, we would not resume it, since the storage
domain did not become invalid.

This patch changes the semantics so storage domain becomes invalid if
the checker is blocked for more than check internal. When the next check
succeeds, we will try to resume VMs paused while the checker was
blocked.

Together with the new multipath configuration, this decreases the chance
of missing a short storage outage, since we now treat any timeout as an
error, even if the check completes successfully after the timeout.

Change-Id: Id5e9e95fdc82cf8074fbadabb615f6baad61bef0
Bug-Url:https://bugzilla.redhat.com/1335176
Signed-off-by: Nir Soffer <nsoffer@redhat.com>