sos-plugin: Add lvm configuration when executing lvm commands
Add lvm configuration when executing lvm command to gather information.
This is important because, for example, lvmetada could be enabled and we
want to disable it when creating the sos report, in addition and as
recommended, the user can define devices filter and we want to override
that filter to get all relevant devices.
In this patch we provide the following lvm config:

    global {
        locking_type = 0
        use_lvmetad = 0
    }

    devices {
        preferred_names = [ '^/dev/mapper/' ]
        ignore_suspended_devices = 1
        write_cache_state = 0
        disable_after_error_count = 3
        filter = [ 'a|^/dev/mapper/.*|', 'r|.*|' ]
    }

Change-Id: Ie1bf98db42c727906a3fa1e826ad582dd8d8d59e
Bug-Url: https://bugzilla.redhat.com/1474566
Signed-off-by: Ala Hino <ahino@redhat.com>