hsm: Use default host id when pool doesn't exists
We want to enable the possibility to force detach a storage domain
even if the host is not yet connected to the storage pool.

This will be used in a scenario when a user will want to recover
its setup after a disaster, and will try to attach a storage domain
with dirty pool UUID of the old storage pool to a new unintialized
Data Center.
Currently the process fails since the host is not yet connected
to the storage pool and therefore there is no host id which we can
use to take a lock on the storage domain.

That is why we should enable the option for force detach a storage
domain and clean its SP_UUID before we will want to connect it to
the storage pool.
Therefore we generate a default host id to get a lock on the storage
domain and detach it from the old storage pool.

Change-Id: I2cf6b329f4f7b461b2bfe77d63e9022d07680e70
Bug-Url: https://bugzilla.redhat.com/1205739
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>