vm: metadata: recover the storage config earlier
To avoid races with the periodic tasks, we restore the
storage device configuration as soon as possible, in Vm.__init__.

Previously, we tried to move the restoration out of already overcrowded
__init__, but doing the restoration in Vm._run opens the gate
for races on startup. To fix them, we would need synchronization
between threads, so moving the restoration in __init__ seems
the (slightly) lesser evil.

Change-Id: I5e025801564a988748a8cf54ab5ebede03b9007a
Signed-off-by: Francesco Romani <fromani@redhat.com>