vdsm tool configurators: bond_defaults - do not depend on time
The bond defaults dump is considered updated if it exists (as it is
created per boot, and its content is dependent on the kernel).

The existing time check (verifying if the period elapsed from the file
update is no older then the time elapsed from the system boot) is no
longer needed, as the dump is placed under /var/run which is assured
to be cleaned on boot.
This patch removes the time checks which are not valid when the time is
updated during the boot.

Change-Id: Ia32455d36d600410a0b792f4e7045b97a08afc70
Bug-Url: https://bugzilla.redhat.com/1482014
Signed-off-by: Edward Haas <edwardh@redhat.com>