net tests func: Fix IPv4 default route assertion
The request flag of defaultRoute is not directly related to the net caps
property of ipv4defaultroute.
ipv4defaultroute is true when the network gateway is also the host
default route.
However, a request with defaultRoute set and no gateway is valid,
resulting with a false ipv4defaultroute.

This patch updates the assertion check to be in sync with the
described logic.

Change-Id: I66ee5173261c84845b325b0b0df3fc91aaf02510
Signed-off-by: Edward Haas <edwardh@redhat.com>