virt: add tunable to disable legacy configuration
The storage devices are the most complex devices to be properly
configured, so when switching to domain XML, we agreed to let them
still use the vm.conf dictionary for the setup.

Furthermore, in order to use the domain XML for the recovery,
the storage devices require a lot of data to be added to the metadata.

To move forward, we add a new tunable, disabled by default, to
enable the storage devices to use the domain XML and the fewest
possible metadata to configure themselves, much like the rest of the
devices.

This will allow testing of the full domain XML flow while Vdsm
still keep working as usual.

Change-Id: Ie5b7178962207e590e31d3ef6d1cc4896188a992
Signed-off-by: Francesco Romani <fromani@redhat.com>