virt: metadata: support disk replica
During the live storage migration, Vdsm dynamically
adds the 'diskReplicate' key to a drive configuration.
This is, for all intents and purposes, a nested drive
configuration. The key will be updated according
to the need of the LSM, and removed once completed.

We need to support this key in order to properly save
and restore the drive configuration.

In the long term, we will still need to save some
information in the metadata, but we will be able
to get rid of the nested drives once the proper
XML support for drives is merged.

Change-Id: I22b016405d2e0384794ad5c6d2377fb9e4e2916b
Signed-off-by: Francesco Romani <fromani@redhat.com>