qga: add QEMU-GA polling facility
We want to replace our oVirt Guest Agent (OGA) with QEMU Guest Agent
(QGA). OGA was pro-active and sent periodicaly information to VDSM. QGA
does not do that and we have to poll for the information repeatedly on
regular basis.

Pool of worker threads is created and each VM is served by a task. The
concept is much like in the periodic.py, in fact the core objects in
executor.py and periodic.py are reused. Special executor managing the
pool of workers is created. Intentionaly it is not the same as for other
periodic operations because there are several reasons (unresponsive
guest agents, OS under load, ...) for workers to block. This would put
other periodic tasks at risk due to lack of available executor threads.

There are several notable risks and issues connected with using QEMU
Guest Agent:

- We (vdsm) have no easy way of knowing whether agent is installed,
  running and responsive. Libvirt has pretty good idea whether agent is
  running or not because it watches QEMU events related to the socket.
  But we have no way of finding that out.

- We can make a call to check the responsiveness of the agent and it
  would quickly return if the agent is not running. However, if the
  agent is running but is not responding then the call will block. This
  is problematic because we cannot pro-actively "ping" the agent.
  Why? Because of the following...

- When we send call to the agent libvirt will lock the domain for the
  time of the operations. This means that lots of other calls to libvirt
  on that domain will block or become unavailable. Notably this affects
  most (if not all) of the operations in periodic.py. All the
  operations handled by VmDispatcher will become unavailable
  (Vm.isDomainReadyForCommands() returns False). Bulk stats will
  probably be delayed.

When gather information we want (at least for now) to prefer information
from OGA over the information from QGA if OGA is available.

Change-Id: I83ea6874f2804631b8be48e166025bc937add2ad
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>