daemonAdapter: Add config option for LIBVIRT_DEBUG
When troubleshooting problems with libvirt it is sometimes beneficial to
enable libvirt debug messages. We already control LIBVIRT_LOG_FILTERS
and LIBVIRT_LOG_OUTPUTS, adding also LIBVIRT_DEBUG only completes the
set.

Configuration option is added that controls setting of LIBVIRT_DEBUG.
This environemnt variable controls printing of client side libvirt
messages. The meaningful values are:

 * 1 or "debug": asking the library to log every message emitted,
   though the filters can be used to avoid filling up the output
 * 2 or "info": log all non-debugging information
 * 3 or "warn": log warnings and errors, that's the default value
 * 4 or "error": log only error messages

Setting LIBVIRT_DEBUG= is the same as unset. If you specify an invalid
value, it will be ignored with a warning.

Change-Id: I62ab6af062bad6fc7acf40665b5c8a60bcf94188
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>