stomp: handle incoming heartbeats
According to stomp specification, Heart-beating can optionally be used
to test the healthiness of the underlying TCP connection and to make sure
that the remote end is alive and kicking. [1]
until now, we ignored incoming heartbeats and raised ValueError
when their value was not 0.
This patch gets the incoming heartbeats and updates the
dispatcher according to their value.
incoming heartbeat will be updated every time a new frame will be
received, indicating connection is still active.
If timeout has exceeded, the frame handler (client or server) will handle
it according to its implementation.

[1] https://stomp.github.io/stomp-specification-1.2.html#Heart-beating

Change-Id: Ib9558ecddef3b57d37cfe4257c1c0ab4f2015064
Signed-off-by: Irit Goihman <igoihman@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1376843