virt: metadata: don't (re)store drive.volumeInfo
We used to store drive.volumeInfo in the metadata,
like we currently do for vm.conf. Turns out, however, that
drive.volumeInfo is no longer used inside Vdsm,
it was one early attempt to store the data that we now
provide inside the drive.volumeChain.
So we can elide this information.

This patch skips this key on both save and restore the drive
configuration to metadata. We will use the Engine XML - and the
recovery code from XML, only in clusterLevel >= 4.2, and
nothing changes in the vm.conf flow,
so backward compatibility is ensured.

With this patch, we also fix the attached rhbz#1481165, which
was caused by the fact that for network (gluster) drives,
volumeInfo contains the unexpected 'hosts' entry, which the
current code is not prepared to deal with.

Change-Id: I0d6f8c5b66a6210e5405c9ad2aac0148dab8f123
Bug-Url: https://bugzilla.redhat.com/1481165
Signed-off-by: Francesco Romani <fromani@redhat.com>