vm: allow erasing of metadata key
A bug in Vm._add_disk_legacy_conf_to_metadata() prevented
the code to remove keys from the drive config stored in the metadata.
Adding and updating keys was fine.
This is relevant when finalizing a live storage migration operation,
because diskReplicateFinish wants to remove the replica information
from the parent drive.

This patch fixes the bug and add unit tests to cover the methods
in the Vm class.

Change-Id: I723fb54faddb2c51f9a6ab6d5bebd577d8b4701e
Signed-off-by: Francesco Romani <fromani@redhat.com>