virt: devices: use 'name' to get drive metadata
The 'vmdevices.common.get_drive_conf_identifying_attrs' provides
the attributes to use to match a device with its metadata.

When the client code needs to update the drive configuration stored
in the metadata, it can invoke this function on the configuration and
use the return value as argument for metadata.Descriptor.device()
to eventually access the metadata.

In the device configuration, the 'name' key carries the exact
same information of the ('iface', 'index') pair. Both uniquely
represent a drive from the guest perspective - e.g. (scsi, 0)
will always be 'sda'.

Problem is the matching code uses the given tuple verbatim,
so ('sda',) is different from ('iface', 0,), even though this
is actually the same drive.

To fix this issue, the function must normalize the parameters.
We choose to always use the name, leveraging the storage.makeName
helper.

One noteworthy drawback is that we now hide the fact that sometimes
the drive configuration is not normalized: drive configuration is
expected to always contain the keys 'iface', 'index', and 'name', the
latter two assigned by Vdsm.

Change-Id: I1377bd8f2cd4677386fe49ddcd3a738d9ed6f35d
Signed-off-by: Francesco Romani <fromani@redhat.com>