vm: get launchPaused flag from metadata
When Engine starts a VM using the domain XML, it will encode
send the 'launch paused' flag in the domain metadata.

This patch makes Vdsm able to consume the value.
Vdsm still honour the flag if present in the vm.conf dictionary,
but only if no domain XML is given.

Change-Id: I620b611c468891e997a12191db7e369e970b27af
Signed-off-by: Francesco Romani <fromani@redhat.com>