virt: metadata: introduce make_element()
The support of the storage device demands complex
layout, not planned in the original design.
Even though the support for storage device is temporary,
we should support them correctly for the time being.

The current code make use of ET.Element() and ET.SubElement()
this is not optimal, but still correct, because the code
doesn't take into account the namespace prefixing.

This is the responsability of the metadata.Metadata() class.
To fill this gap, we introduce the make_element() helper,
and we change the _dump_device helpers to use it.

Change-Id: Id148e1329b49e16ccd5fb8d9689708562d062d36
Signed-off-by: Francesco Romani <fromani@redhat.com>