blockSD: Rename _extendlock
_extendlock is used is used to protect the metadata mapping, which may
change either when lvs are created, removed, resized or when pvs are
added, removed, resized, when using old storage domain versions.
In this patch, we rename the lock to metadata_lock to better reflect its
usage, and we declare it public as it will be used outside of this
class, e.g. commit Ia9f823 (api: Implement reduceVolume verb).

Change-Id: I2bf0d400df682855676c65dabac7f3c0c0adc88f
Signed-off-by: Ala Hino <ahino@redhat.com>