API: Don't log tracebacks on non-existent VMs
When a Vdsm operation is requested on a VM that doesn't exist, we log an
exception with a traceback.  This pollutes Vdsm log with tracebacks even in
some regular operations such as powering off a VM, see the referenced bug.

Calling an operation on a VM that doesn't exist is not a Vdsm error and the
traceback produced in API.py is useless.  So let's mark that exception as
expected.

Change-Id: I9ad2deb39c54c5f4c36dc0006998536e21429676
Bug-Url: https://bugzilla.redhat.com/1478712
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>