virt: devices: tunable to disable the XML refresh
Some device classes may or may not need to refresh the XML configuration
that Engine sends. Add a tunable to easily enable/disable XML refresh on
a per-device-class basis.

This tunable is under the `devel' section, and should be used when tuning
the Engine XML support. Should never be used in production environment
(as any other in the `devel' section), and we may likely go away as soon
as we have a stabilized the engine XML flow.

Change-Id: I0c3f8c8494e5ac5772bfb94e407ba453f2f9ee2c
Signed-off-by: Francesco Romani <fromani@redhat.com>