net: Fail early on initialization of LLDP
If LLDP is detected as not active, the initialization of LLDP
is not continued.
This has the advantage that there is no further time spent on
the initialization of LLDP, if LLDP is not working properly.

Change-Id: I6d0ca7613b7f3cf022ed46d07eea615367af36b6
Bug-Url: https://bugzilla.redhat.com/1488358
Signed-off-by: Dominik Holler <dholler@redhat.com>
(cherry picked from commit f2e9f7c2dcebd105c5db5457a39b31fec34bfdb0)