virt: Destroy libvirt domain when destroying incoming migration
When Destroy is called on migration destination we just set the VM to Down
status and don't destroy the libvirt domain.  It's OK in typical cases since
the migration is canceled on the source and the domain is terminated by
libvirt.

But there is at least one situation where it's not sufficient.  When Engine
loses connection to the source host during migration and wants to cancel the
migration it calls Destroy on the destination host.  Then the VM is removed
from Vdsm but the migration may still be running and eventually finish,
resulting in the VM running on the destination without Vdsm and Engine being
aware about it.  Split brain follows sooner or later.

So when destroy is requested on the destination we try to destroy the libvirt
domain.  The call may not succeed if the domain has already been destroyed or
it's not yet started on the destination.  In the latter case there is still
theoretical risk of split brain but with much lower probability than without
this patch.

Change-Id: Id5532fd5322d03f9a42dd16ff00bee0e3f3c9d94
Bug-Url: https://bugzilla.redhat.com/1486543
Backport-To: 4.1
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>