virt: Save disk metadata once devices are prepared
Once disks are prepared we should save their metadata as soon as possible to
not lose that configuration on contingent Vdsm restart or crash.  Since the VM
is not set up or running yet, we create the most simplistic VM domain to store
the metadata too.  It is going to be updated later on real VM start.

A notable case is incoming migration, where the domain may not be yet available
at the time metadata is stored.  So for incoming migrations we store metadata
only after we attach to the domain.

This patch restores disk data saving, through a different mechanism, that used
to be performed when recovery files were used.

Change-Id: I297d3d0337f59bf5fb3b9e638698d234c27e3de5
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>