virt: metadata: per-device custom properties
The per-device metadata section of the domain XML
contains any device data which doesn't properly fit
elsewhere in the domain XML. This include oVirt-specific
data.

Except for storage devices, this section is flat, with
no nesting allowed. This is because the metadata is
supposed to be used sparingly, because most if not all
the configuration a domain may need should find its place
in the domain XML proper.

Unfortunately, the custom properties badly fit in this
assumption. The device custom properties are a bunch of
key-value pairs, and the the network subsystem makes heavy
use of them, partially because key functionality is implemented
in hooks, which in turn require (or are activated by) per-device
custom properties.

To better fit this case, we allow one nested section
in the per-device metadata, called "custom", to hold
the key-value custom properties. No further nesting is allowed.
Engine can trivially translate the old custom properties
to the new XML format, and with this patch Vdsm can consume
them - and also encode them should the need arise.

Example:

  <metadata>
    <ovirt-vm:vm>
      <ovirt-vm:device mac_address='${MAC}'>
        <ovirt-vm:custom>
          <ovirt-vm:vnic_id>${UUID}</ovirt-vm:vnic_id>
	  <ovirt-vm:provider_type>EXTERNAL_NETWORK</ovirt-vm:provider_type>
        </ovirt-vm:custom>
      </ovirt-vm:device>
    </ovirt-vm:vm>

This patch also addresses the attached BZ.
In the attached BZ, the problem is that the hooks are not triggered,
because in the domain XML flow that field is not populated, because
Vdsm doesn't know where to look for the per-device custom properties[1].

Please note that in order to fix the attached BZ, Engine must also
encode the device custom properties in the metadata section, and
send them in the domain XML.

+++

[1] Vdsm already parse some well-known XML fields to populate some
custom properties when the mapping is not ambiguous, like 'queues',
but none is relevant on this case.

Change-Id: I81f59632fd81390b5df1e5f8dc2e85818ac6597a
Bug-Url: https://bugzilla.redhat.com/1483309
Signed-off-by: Francesco Romani <fromani@redhat.com>