network: skip ifdown if bridge does not exist
Until now we called ifdown during setup of a bridge even when
it did not exist. That led to performance drop during setup of
high number of new networks.

With this patch we don't call ifdown if a bridge is not found
in the system.

Setup of 26 new bridged networks without this patch takes
15 seconds, with this patch it only takes 5.

Change-Id: If32f575d11ad7ad72911e25ef7a21756d8fdff55
Signed-off-by: Petr Horacek <phoracek@redhat.com>