network: remove vlan directly if not a top device
In case vlan is not a top device of a network we can remove it
directly via `ip link delete` instead of calling `ifdown` on it.
This is possible because of it does not keep any L3 configuration
that would need a cleanup.

Time spent on setupNetworks call of 51 networks:
         add  edit  remove
master:  47s  44s   28s
patched: 47s  33s   17s

Change-Id: I51611c38862e73854d91364aea073888c761c53c
Signed-off-by: Petr Horacek <phoracek@redhat.com>