virt: storage: introduce the base26 module
Linux names the device using their base26-encoded drive
number. For example, on any given bus (SCSI, virtio...)
the 0th device name is 'a', the 1st is 'b', and so
forth up to 'zzz' - this limit is in the kernel code.

This patch extracts the base26 encoding code - currently
broken, see https://gerrit.ovirt.org/#/c/81128/ and moves
it into the newly introduced 'base26.py' module to
improve the readability and to make room to the decoding
code, which will be added by the followup patch.

Only code movement, no changes.

Change-Id: I7f809256bf9514ab3bb84f50416ca1e2c597f130
Signed-off-by: Francesco Romani <fromani@redhat.com>