lib: make base26.encode() work like Linux
This patch changes the implementation of
base26.encode() to behave like the linux kernel.

The original implementation of base26.encode() uses
a different algorithm than the Linux kernel.
Thus, the names Vdsm produces are "wrong", meaning
that, past the 26-th disk device, they are different
from the one Linux uses.

This breaks the assertion that
device == drivename.make(*drivename.split(device))

but other than that is expected to have near zero-impact
on existing VMs.
The only possible source of troubles is if we migrate
one VM from buggy Vdsm to fixed Vdsm. If the VM has a very
complex disk configurations with holes in the sequence,
(e.g. 27 disks attached, #28 hotunplugged, #29 attached),
it is theoretically possible that hotunplug will fail
because the new base26 code will pick a correct but conflicting
device name, depending on the device index.

But the only drawback would be hotplug fail, and a trivial
workaround (in one already extreme scenario) is just to pick
a different bus (e.g. SCSI vs virtio).

Change-Id: Ib83c15f4964db59f4a4ae0d5aaee481759360197
Signed-off-by: Francesco Romani <fromani@redhat.com>