imageio: expect bytes as http response type
imagetickets_test uses a FakeResponse object to mock the http response
that vdsm gets from the daemon.
Since reading from an http socket results in bytes and not unicode,
there's no reason to work with unicode in the test infrastructure.
Therefore, this patch changes the type of the data field of the
FakeResponse object to be bytes rather than unicode.

In imageticket.py._read_content, this patch decodes the response using
utf8, what makes it compatible with Python3.

Change-Id: Ie887fe25513e17ae1b7d1bc1f0d4bb8651a62a3f
Signed-off-by: Idan Shaby <ishaby@redhat.com>