virt: do not override devices
Storage device used to not support the XML-based initialization.
We added a safety net to fetch them from the old vm.conf
but this is not needed anymore, and it is actually hiding bugs for
the Engine-XML flow.

We already have a conditional to fully enable the engine-XML flow,
but this hunk was no included by mistake. This patch fixes that.

Change-Id: I792094d736ecafe134b3dd2838e6d848979fc1c2
Signed-off-by: Francesco Romani <fromani@redhat.com>