virt: storage: disable from_xml_tree
When using the engine XML flow, the storage device need
special two-phase processing, because their parameters
need to be updated on per-host basis.

For pratical purposes, we do this setup after the parsing
of the XML snippet (augmented with metadata) to restore
the storage.Drive params, but before to create the
storage.Drive instance; this way we complicate the
flow a bit, but we minimize the overall changes[1].

This patch prepares the path disabling the Drive.from_xml_tree
method, to make sure storage devices are not processed
twice.

+++

[1] The setup code include two major steps: path preparation
and the running of the hooks after the image has been prepared.
Both stages expect to deal with update exposing the dict interface,
which storage.Drive implements only partially.
Since this ambiguity between object interface (obj.foo vs obj['foo'])
is already confusing, I prefer to differentiate the storage
device processing a bit.

Change-Id: I31ce3fc62ab3fb3797d08eb1bbe7db077b85b2e5
Signed-off-by: Francesco Romani <fromani@redhat.com>