fileSD: Gracefully handle purgeImage delete race
When deleting file-based images we first call deleteImage to mark the
image for deletion (by renaming the directory with a known prefix).  We
complete the deletion by calling purgeImage which first removes the
contents of the deleted image directory and then the directory itself.

If the sdCache is invalidated between the calls to deleteImage and
purgeImage, then the FileSD.imageGarbageCollector method will be invoked
when the FileSD is recreated.  This will remove the deleted image
directory early.  In this case, purgeImage is failing when the directory
cannot be deleted.

We already gracefully handle missing volume files by logging a warning.
We should do the same when the directory itself is missing.

Change-Id: If19058d23abd088826b0877bc541bfc687440b22
Bug-Url: https://bugzilla.redhat.com/1342550
Bug-Url: https://bugzilla.redhat.com/1502213
Signed-off-by: Adam Litke <alitke@redhat.com>