numa: report memory stats even without NUMA
If libvirt is built without numactl support, it will report a "fake"
single cell numa topology. However it will fail to report node memory
statistics for the numa cell 0 with an error message.

This prevents ovirt-engine from correctly displaying the host's status.

To work around this issue, memory statistics for a single cell host
are retrieved for the entire host which works even without numactl
support. The behavior doesn't change.

Change-Id: Iee90206d46cb726d28a923f8c49c28bb50a49533
Signed-off-by: Viktor Mihajlovski <mihajlov@linux.vnet.ibm.com>