gencerts: use yesterday as the valid date for the certificate
Following the same pattern as certs for engine, use yesterday's
date as the starting date for certificate validity, in case the
local hardware clock is not UTC (or will otherwise be drift
corrected).

Even though we recommend ntp/chrony, it's possible that this will
not always be done before vdsm-tool configure, so it's better to
be safe.

Change-Id: I6ae11915dc4f76113702e5104073f55e3f41ac6b
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1474213
Signed-off-by: Ryan Barry <rbarry@redhat.com>