net: Wait for link-up while query netinfo nic speed
During the network caps data collection, a nic speed may be query while
the link is still down (due to a previous setup action).
A max of 2 seconds wait period is added, giving the links a chance to
recover.

On the first link that is detected as down, the 2sec period is
initiated. This is an one per report wait period (and not per link).

Change-Id: I25ba2b373235fb040a43ce2bfd5343243eed47f3
Bug-Url: https://bugzilla.redhat.com/980363
Signed-off-by: Edward Haas <edwardh@redhat.com>