vm: skip readonly drives in getChunkedDrives()
Strictly speaking, a readonly chunked drive makes sense and it
is a supported configuration. But we use (and we should use)
Vm.getChunkedDrives() only in the drive extensions flow,
and in this flow we don't care about readonly drives - those are
obviously never extended online.

Thus, this patch makes getChunkedDrives() skip readonly drives.

Change-Id: I07c43cc278b858ac79f7f6ee1ac5fd695312b07c
Signed-off-by: Francesco Romani <fromani@redhat.com>