virt: trick OVS display network
A libvirt network with type="bridge" and bridge="ovs_nb_interface"
can be used as a network for libvirt graphics with autodetected IP
address. Note that this network won't be usable for VM interfaces
since it is not a real bridge.

By doing this trick we can get rid of some special handling of
IP resolution and part of migration hook.

Change-Id: I967dbae23861eb63cfc7b95bb0e62da625f778c1
Signed-off-by: Petr Horacek <phoracek@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1400952