net: Exclude nic speed info from netinfo and add it to netcaps only
NIC speed information was part of netinfo report, which is frequently
used during the setup flows.
As the speed is not required during the setup flow and has value only in
the network capability report, speed is no longer part of netinfo.

The NIC speed info is now added only as part of netcaps.

Change-Id: I3bf091dd99d802e36ad5d51eaa41d58773039caa
Signed-off-by: Edward Haas <edwardh@redhat.com>