virt: storage: use 'devtype' as matching attribute
The metadata handling code needs some attributes
to match a device instance with its metadata.
Storage devices used the (type, name) tuple, but
the "type" attribute clash with the type hint the
generic metadata code uses.

We switch to the (devtype, name) tuple to avoid this clash.
The 'name' attribute alone is sufficient to uniquely match
a device, but we keep the newly renamed 'devtype' attribute
to make the code more future-proof.

This bug didn't surface before because was hidden by another
bug[1], and because we overlooked a quirk in how the ElementTree
parser works, that furterly prevented this bug to emerge[2]

[1]: https://gerrit.ovirt.org/#/c/82288/
[2]: see the tests of [1] to learn about the quirk

Change-Id: I2d9d73f2ee9f51223c282c66ed11743edd2eb983
Signed-off-by: Francesco Romani <fromani@redhat.com>