virt: watermark: introduce the drivemonitor module
This patch introduces the drivemonitor code, which
will hold all the logic needed to set and clear
the block threshold for drives, and the code
to handle the block threshold events.

The next patches will build on top of this code
to use and consume the libvirt BLOCK_THRESHOLD event.

Change-Id: I3ca9604cd1c0172954ec7a6a424def7a1093c0d0
Signed-off-by: Francesco Romani <fromani@redhat.com>