net tests: Relocate testReorderBondingOptions to the new func tests
The tests has mainly verified that if the mode is specified last in
the bond options, it is still handled first (the other options may
have no meaning before the mode switch).

In addition, the original test used named values in the bond options,
which requires special normalization in the tests assertion.
(name to numeric translation)

The network part of the original test has been removed, as it only
relates to bonds.

Change-Id: I816d7a68984de95ee4ae4e65a7a34124bb26bc45
Signed-off-by: Edward Haas <edwardh@redhat.com>