net: An unowned existing bridge should be ignored
The ifcfg configurator supports the ability to reuse an existing bridge
without tearing it down. However, an unowned bridge (either with an ifcfg
file or without) should not be considered for this case, as it may require
redefinition.

And old functional test (testSetupNetworksOverExistingBridge) was used
to check a scenario where a bridge existed without an ifcfg file,
however, the test was broken and never really created the bridge.
This patch relocated the test under the new functional tests and fixes
the production code to ignore unowned bridges.

Change-Id: I4f4db6190c5bc633bb39a5ca4fe85bf7bbc4de77
Signed-off-by: Edward Haas <edwardh@redhat.com>