virt: Auto-resume VMs only conditionally
When a VM experiences an I/O error, it gets paused.  Once the problem is gone,
Vdsm may resume the VM automatically.  This is however not always desirable, so
we make this behavior configurable.  Engine can set auto-resume behavior in
metadata `resumeBehavior' item now, to one of the following values:

- auto_resume: The VM is resumed.
- leave_paused: The VM will remain paused.
- kill: The VM will be killed if it is stopped for too long; otherwise it is
  resumed.

If the setting is not present, we keep the current behavior, i.e. the VM is
resumed.

This patch implements only `auto_resume' and `leave_paused' cases, `kill' will
be implemented in next patches.

Change-Id: Ib88725911d92e543674de26b06ef58e33cc74b87
Bug-Url: https://bugzilla.redhat.com/1317450
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>