virt: Track pause time
We would like to kill, depending on autoResume setting in metadata, VMs that
are paused due to an I/O error for too long time.  In order to be able to do
that, we must track the time when the VM was paused.  This patch implements the
tracking and storing it to metadata.

Change-Id: I73a7286f5ba849b95a75ccde453699d954928c6f
Bug-Url: https://bugzilla.redhat.com/1317450
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>