virt: Kill VMs paused for too long
If a VM is paused due to an I/O error for a defined interval and its
auto-resume setting is `kill', it should be killed rather than resumed.  The
interval is set similar to sanlock lease expiration interval, 80s.  It needs to
be more than 0 to tolerate short time hiccups and must not be longer than time
it takes Engine to move VM to Unknown status, which would trigger a restart
elsewhere (~5 minutes).

Change-Id: I5e74f7a828e3b7fbfcf774fc7c4304d587a4b278
Bug-Url: https://bugzilla.redhat.com/1317450
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>