lvm: Add stderr to CannotCreateLogicalVolume
When a creation of a logical volume (LV) fails an exception is raised,
as expected, but this exception isn't very useful unfortunately.
While it includes the IDs required to identify the failed LV creation,
it doesn't contain any information about the error that occurred.

This patch adds the stderr output to the exception so it can be
logged.

Change-Id: Ib6a6a62df2f6dce512c3c18f8cf0a48c721f89f7
Related-To: https://bugzilla.redhat.com/1497931
Signed-off-by: Allon Mureinik <amureini@redhat.com>