spec: Fix kernel dependency (removing trailing arch)
58ebca9 introduced a dependency on a newer kernel version, however it
included the kernel arch.

When trying to install vdsm on a host that has kernel
version = 3.10.0-693.2.1.el7, the installation fails because from RPM
perspective 3.10.0-693.2.1.el7 < 3.10.0-693.2.1.el7.x86.64.
Nevertheless, the version of the 'kernel' package doesn't include the
'x86_64' suffix.

The error doesn't occur in check-merged since the kernel version in the
centos7.4 image is greater than 3.10.0-693.2.1.el7.

Change-Id: Icd4d31f622404fea8587d3a908c17267b49e2416
Signed-off-by: Edward Haas <edwardh@redhat.com>