virt: metadata: preserve deeply-nested metadata
The metadata module initially didn't plan to support more than
one level than nesting. It was not supported, for example, to
store data structure (list or dicts), only plain data like
strings or numbers.

This limitation was artificial and intentional, to discourage
the consumers to abuse the metadata, and to discourage bloat in
this area.

The need to support storage drives configuration changed the
landscape, so support for deeply nested data structure was added.
A number of places assumed no nesting, and not everyone was fixed.
This patch fixes the remaining places and adds a test to cover
the last one we are aware of.

Please note that even though we plan to deprecate the support
for storage device in metadata, the payload devices will be
kept in the metadata, and they will likely require some
deeper nesting.

Change-Id: I8f96193bc28f01df60f282943b749ea5809a1e4f
Signed-off-by: Francesco Romani <fromani@redhat.com>