virt: powerdown: fix qemuGuestAgent usage
The Vm.qemuGuestAgent* helpers correctly raise exceptions
on failure, instead of reporting errors using the response module.
This is correct because the creation of responses belongs to the API
layer, not to the application layer.

Unfortunately, the clients in the vmpowerdown module were not updated,
expecting responses as return value.

This patch fixes the vmpowerdown module to catch the exceptions
in case of failures, and act accordingly.

Change-Id: I061ec5f712f9acb2a1098893775259e961c7df61
Bug-Url: https://bugzilla.redhat.com/1487064
Signed-off-by: Francesco Romani <fromani@redhat.com>