vm: fix Vm.get_balloon_info
At steady state, we will have exactly one balloon device,
one 'real' one if Engine asked Vdsm so, or one bogus otherwise.

Unfortunately, this is not true while the VM is starting up.
Nothing prevents the getStats() flow to be called in this case,
and if so it will fail, emitting one scary and unneeded stacktrace
in the logs.

This patch makes Vm.get_balloon_info() behave correctly when we
have no Balloon Devices.

Change-Id: I42708f5550cd69705705388afd19a4ac91644485
Signed-off-by: Francesco Romani <fromani@redhat.com>