spec: Require libvirt-python fix
Bug 1496517 (solved for EL 7.4.z by bug 1497197) describes an issue
where libvirt's python bindings do not properly release the GIL in
virDomain.getMemoryStats and virDomain.diskErrors.
If either of these methods are called on a VM that uses a disk placed
on non-responsive storage, vdsm itself may get blocked.

This patch requires libvirt-python-3.2.0-3.el7_4.1, which provides a
fix for the aforementioned issue.

Change-Id: Ifac89552d03ff3e3de7436e039c3e4e4fe82a63e
Bug-Url: https://bugzilla.redhat.com/1505280
Signed-off-by: Allon Mureinik <amureini@redhat.com>