vm: add support to force the resume behavior
This patch extends Vm.maybe_resume() to accept
the ResumeBehavior as optional parameter. If None is specified, the Vm
setting received through the domain metadata will be used.

This is useful because Vm.maybe_resume() may be called in the Vm internal
flows, where stopping - or worse killing a VM may lead to very
surprising behaviour.

Change-Id: Id3f80a82c8ef45cc2173a0d7e2967a318649d5fa
Related-To: https://bugzilla.redhat.com/1317450
Signed-off-by: Francesco Romani <fromani@redhat.com>