qemuimg: Handle leaked clusters when running qemuimg check
We use qemuimg check to get the volume offset in order to reduce its
size to optimal. However, if the base volume have leaked clusters,
qemuimg check return the info in stdout, but the return code is 3
(our code expects 0 as the success return code). As the out is
still returned even if there leaked clusters, in this patch we
handle the error code and return the expected result.

If leaked cluster found, following message is logged (message was cut
for a nice representation):
WARN [QemuImg] 223 leaked clusters found on the image

Change-Id: Ie8314deaaad33f8a5b40ae569cd445d674d29017
Bug-Url: https://bugzilla.redhat.com/1506503
Signed-off-by: Ala Hino <ahino@redhat.com>