supervdsm: mkimage: remove floppy if existing
Under normal circumstances we clean leftover payload files when we
destroy VM. Due to unclean host shutdowns and possibly bugs in VDSM an
image may be left in the payload directory. This is problematic in case
of floppy images that are created using mkfs.msdosfs because it refuses
to overwrite existing image. Such image then blocks start of the VM and
manual intervention from administrator is required.

There appears to be no way to force mkfs.msdosfs to overwrite the
existing image so we have to do the dirty job.

Change-Id: I7ddb24ba472d797ba7bce99da7cd11c32efd5886
Bug-Url: https://bugzilla.redhat.com/1506517
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>