virt: Always resume VMs paused due to ENOSPC
We introduced conditional resume of a paused VM in previous patches.  However
the resume shouldn't be conditional when we resume a VM stopped due to ENOSPC,
i.e. when disk extension is in action.

With this patch we revert the misplaced maybe_resume calls.  We call
maybe_resume only in storage domain state change callback now.

Change-Id: I7bec3d6d21948028504acd2162e0eb9049a810cc
Bug-Url: https://bugzilla.redhat.com/1317450
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>