net: save mapping between networks and ovs bridges
external_ids:ovn-bridge-mappings is used by OVN to map between network
name and OVS bridge on which it is set on. In case OVN has a localnet port
assigned to a network, this information is used to find the correct bridge
for the creation of a patch.

This information is always saved after setupNetworks of OVS networks.

Please note that we do not cover this side effect in any current test.
It should be covered by OVN related tests as system test. But we may want
to add coverage internally as well.

Change-Id: I24ee411be4abf469307cb9e22fe7d61e272412e9
Signed-off-by: Petr Horacek <phoracek@redhat.com>