virt: normalize storage devices from XML
Vdsm tracks the drive indexes per-bus, in order to be
able to find a valid device name (from device interface and index)
when a new disk is hotplugged.

This tracking done when a VM is started using the old vm.conf format
(oVirt <= 4.1), but not yet when a VM is started using the domain XML
sent by the client.

This leads to duplicate device names when a disk is hotplugged, which
will make the operation fail.

This patch makes sure that the internal tracking data structure is
properly initialized even if a VM is started from the domain XML.

Change-Id: Ic9b90055df8d20d75637b7bc6999701771d2f0f7
Signed-off-by: Francesco Romani <fromani@redhat.com>